requirements.txt
cheroot
flask
flask_wtf
requests
beautifulsoup4
lxml
Pillow
click
Flask-Squeeze
Flask-Minify